chore: remove deprecated client endpoints
Fixes gravitee-io/issues#5475
