+++ gravitee-am-gateway/gravitee-am-gateway-handler/gravitee-am-gateway-handler-common/src/main/java/io/gravitee/am/gateway/handler/common/spring/CommonConfiguration.java
@@ -123,7 +123,7 @@ public class CommonConfiguration {
     }
 
     @Bean
-    public ClientSyncService clientService() {
+    public ClientSyncService clientSyncService() {
         return new ClientSyncServiceImpl();
     }
