+++ gravitee-am-gateway/gravitee-am-gateway-handler/gravitee-am-gateway-handler-oidc/src/main/java/io/gravitee/am/gateway/handler/oidc/service/clientregistration/impl/ClientServiceImpl.java
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.am.service.impl;
+package io.gravitee.am.gateway.handler.oidc.service.clientregistration.impl;
 
 import io.gravitee.am.common.oauth2.GrantType;
 import io.gravitee.am.common.utils.RandomString;
 import io.gravitee.am.common.utils.SecureRandomString;
+import io.gravitee.am.gateway.handler.oidc.service.clientregistration.ClientService;
 import io.gravitee.am.identityprovider.api.User;
 import io.gravitee.am.model.Application;
 import io.gravitee.am.model.account.AccountSettings;
@@ -25,17 +26,9 @@ import io.gravitee.am.model.application.ApplicationAdvancedSettings;
 import io.gravitee.am.model.application.ApplicationOAuthSettings;
 import io.gravitee.am.model.application.ApplicationSettings;
 import io.gravitee.am.model.application.ApplicationType;
-import io.gravitee.am.model.common.Page;
 import io.gravitee.am.model.oidc.Client;
 import io.gravitee.am.service.ApplicationService;
-import io.gravitee.am.service.ClientService;
-import io.gravitee.am.service.exception.ClientNotFoundException;
 import io.gravitee.am.service.exception.InvalidClientMetadataException;
-import io.gravitee.am.service.exception.TechnicalManagementException;
-import io.gravitee.am.service.model.NewClient;
-import io.gravitee.am.service.model.PatchClient;
-import io.gravitee.am.service.model.TopClient;
-import io.gravitee.am.service.model.TotalClient;
 import io.gravitee.am.service.utils.GrantTypeUtils;
 import io.reactivex.Completable;
 import io.reactivex.Maybe;
@@ -45,12 +38,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * @author David BRASSELY (david.brassely at graviteesource.com)
@@ -79,187 +68,6 @@ public class ClientServiceImpl implements ClientService {
                     }
                     return client;
                 });
-    }
-
-    @Override
-    public Maybe<Client> findByDomainAndClientId(String domain, String clientId) {
-        LOGGER.debug("Find client by domain: {} and client id: {}", domain, clientId);
-        return applicationService.findByDomainAndClientId(domain, clientId)
-                .map(Application::toClient)
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find client by domain: {} and client id: {}", domain, clientId, ex);
-                    return Maybe.error(new TechnicalManagementException(
-                            String.format("An error occurs while trying to find client by domain: %s and client id: %s", domain, clientId), ex));
-                });
-    }
-
-    @Override
-    public Single<Set<Client>> findByDomain(String domain) {
-        LOGGER.debug("Find clients by domain : {}", domain);
-        return applicationService.findByDomain(domain)
-                .map(pagedApplications -> pagedApplications
-                        .stream()
-                        .map(Application::toClient)
-                        .collect(Collectors.toSet()));
-    }
-
-    @Override
-    public Single<Set<Client>> search(String domain, String query) {
-        LOGGER.debug("Search clients for domain {} and with query {}", domain, query);
-        return applicationService.search(domain, query, 0, Integer.MAX_VALUE)
-                .map(pagedApplications -> pagedApplications.getData()
-                        .stream()
-                        .map(Application::toClient)
-                        .collect(Collectors.toSet()));
-    }
-
-    @Override
-    public Single<Page<Client>> search(String domain, String query, int page, int size) {
-        LOGGER.debug("Search clients for domain {} and with query {}", domain, query);
-        return applicationService.search(domain, query, page, size)
-                .map(pagedApplications -> {
-                    Set<Client> clients = pagedApplications.getData()
-                            .stream()
-                            .map(Application::toClient)
-                            .collect(Collectors.toSet());
-                    return new Page<>(clients, pagedApplications.getCurrentPage(), pagedApplications.getTotalCount());
-                });
-    }
-
-    @Override
-    public Single<Page<Client>> findByDomain(String domain, int page, int size) {
-        LOGGER.debug("Find clients by domain: {}", domain);
-        return applicationService.findByDomain(domain, page, size)
-                .map(pagedApplications -> {
-                    List<Client> clients = pagedApplications.getData()
-                            .stream()
-                            .map(Application::toClient)
-                            .collect(Collectors.toList());
-                    return new Page<>(clients, pagedApplications.getCurrentPage(), pagedApplications.getTotalCount());
-                })
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find clients by domain: {}", domain, ex);
-                    return Single.error(new TechnicalManagementException(
-                            String.format("An error occurs while trying to find clients by domain: %s", domain), ex));
-                });
-    }
-
-    @Override
-    public Single<Set<Client>> findAll() {
-        LOGGER.debug("Find clients");
-        return applicationService.findAll()
-                .map(pagedApplications -> pagedApplications
-                        .stream()
-                        .map(Application::toClient)
-                        .collect(Collectors.toSet()))
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find clients", ex);
-                    return Single.error(new TechnicalManagementException("An error occurs while trying to find clients", ex));
-                });
-    }
-
-    @Override
-    public Single<Page<Client>> findAll(int page, int size) {
-        LOGGER.debug("Find clients");
-        return applicationService.findAll(page, size)
-                .map(pagedApplications -> {
-                    List<Client> clients = pagedApplications.getData()
-                            .stream()
-                            .map(Application::toClient)
-                            .collect(Collectors.toList());
-                    return new Page<>(clients, pagedApplications.getCurrentPage(), pagedApplications.getTotalCount());
-                })
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find clients", ex);
-                    return Single.error(new TechnicalManagementException("An error occurs while trying to find clients", ex));
-                });
-    }
-
-    @Override
-    public Single<Set<TopClient>> findTopClients() {
-        LOGGER.debug("Find top clients");
-        return applicationService.findTopApplications()
-                .map(topApplications -> {
-                    return topApplications
-                            .stream()
-                            .map(topApplication -> {
-                                TopClient topClient = new TopClient();
-                                topClient.setClient(topApplication.getApplication().toClient());
-                                topClient.setAccessTokens(topApplication.getAccessTokens());
-                                return topClient;
-                            })
-                            .collect(Collectors.toSet());
-                })
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find top clients", ex);
-                    return Single.error(new TechnicalManagementException("An error occurs while trying to find top clients", ex));
-                });
-    }
-
-    @Override
-    public Single<Set<TopClient>> findTopClientsByDomain(String domain) {
-        LOGGER.debug("Find top clients by domain: {}", domain);
-        return applicationService.findTopApplicationsByDomain(domain)
-                .map(topApplications -> {
-                    return topApplications
-                            .stream()
-                            .map(topApplication -> {
-                                TopClient topClient = new TopClient();
-                                topClient.setClient(topApplication.getApplication().toClient());
-                                topClient.setAccessTokens(topApplication.getAccessTokens());
-                                return topClient;
-                            })
-                            .collect(Collectors.toSet());
-                })
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find top clients by domain", ex);
-                    return Single.error(new TechnicalManagementException("An error occurs while trying to find top clients by domain", ex));
-                });
-    }
-
-    @Override
-    public Single<TotalClient> findTotalClientsByDomain(String domain) {
-        LOGGER.debug("Find total clients by domain: {}", domain);
-        return applicationService.countByDomain(domain)
-                .map(totalClients -> {
-                    TotalClient totalClient = new TotalClient();
-                    totalClient.setTotalClients(totalClients);
-                    return totalClient;
-                })
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find total clients by domain: {}", domain, ex);
-                    return Single.error(new TechnicalManagementException(
-                            String.format("An error occurs while trying to find total clients by domain: %s", domain), ex));
-                });
-    }
-
-    @Override
-    public Single<TotalClient> findTotalClients() {
-        LOGGER.debug("Find total client");
-        return applicationService.count()
-                .map(totalClients -> {
-                    TotalClient totalClient = new TotalClient();
-                    totalClient.setTotalClients(totalClients);
-                    return totalClient;
-                })
-                .onErrorResumeNext(ex -> {
-                    LOGGER.error("An error occurs while trying to find total clients", ex);
-                    return Single.error(new TechnicalManagementException("An error occurs while trying to find total clients", ex));
-                });
-    }
-
-    @Override
-    public Single<Client> create(String domain, NewClient newClient, User principal) {
-        LOGGER.debug("Create a new client {} for domain {}", newClient, domain);
-        Client client = new Client();
-        client.setDomain(domain);
-        client.setClientId(newClient.getClientId());
-        client.setClientSecret(newClient.getClientSecret());
-        client.setClientName(newClient.getClientName());
-        //AM UI first step client creation does not provide field to specify redirect_uris, so better no use code grant by default.
-        client.setAuthorizedGrantTypes(Arrays.asList());
-        client.setResponseTypes(Arrays.asList());
-        return create(client);
     }
 
     @Override
@@ -312,14 +120,6 @@ public class ClientServiceImpl implements ClientService {
 
         return applicationService.update(convert(client))
                 .map(Application::toClient);
-    }
-
-    @Override
-    public Single<Client> patch(String domain, String id, PatchClient patchClient, boolean forceNull, User principal) {
-        LOGGER.debug("Patch a client {} for domain {}", id, domain);
-        return findById(id)
-                .switchIfEmpty(Maybe.error(new ClientNotFoundException(id)))
-                .flatMapSingle(toPatch -> this.update(patchClient.patch(toPatch, forceNull)));
     }
 
     @Override
