--- gravitee-am-management-api/gravitee-am-management-api-rest/src/main/java/io/gravitee/am/management/handlers/management/api/resources/organizations/environments/domains/ClientEmailsResource.java
@@ -1,128 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.am.management.handlers.management.api.resources.organizations.environments.domains;
-
-import io.gravitee.am.identityprovider.api.User;
-import io.gravitee.am.management.handlers.management.api.resources.AbstractResource;
-import io.gravitee.am.model.Acl;
-import io.gravitee.am.model.Template;
-import io.gravitee.am.model.permissions.Permission;
-import io.gravitee.am.service.ClientService;
-import io.gravitee.am.service.DomainService;
-import io.gravitee.am.service.EmailTemplateService;
-import io.gravitee.am.service.exception.ClientNotFoundException;
-import io.gravitee.am.service.exception.DomainNotFoundException;
-import io.gravitee.am.service.model.NewEmail;
-import io.gravitee.common.http.HttpStatusCode;
-import io.gravitee.common.http.MediaType;
-import io.reactivex.Maybe;
-import io.swagger.annotations.*;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-import javax.ws.rs.*;
-import javax.ws.rs.container.AsyncResponse;
-import javax.ws.rs.container.ResourceContext;
-import javax.ws.rs.container.Suspended;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.Response;
-import java.net.URI;
-
-/**
- * @author Titouan COMPIEGNE (titouan.compiegne at graviteesource.com)
- * @author GraviteeSource Team
- */
-@Api(tags = {"email"})
-@Deprecated
-public class ClientEmailsResource extends AbstractResource {
-
-    @Autowired
-    private EmailTemplateService emailTemplateService;
-
-    @Autowired
-    private DomainService domainService;
-
-    @Autowired
-    private ClientService clientService;
-
-    @Context
-    private ResourceContext resourceContext;
-
-    @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "Find a email for a client",
-            notes = "User must have APPLICATION_EMAIL_TEMPLATE[READ] permission on the specified client " +
-                    "or APPLICATION_EMAIL_TEMPLATE[READ] permission on the specified domain " +
-                    "or APPLICATION_EMAIL_TEMPLATE[READ] permission on the specified environment " +
-                    "or APPLICATION_EMAIL_TEMPLATE[READ] permission on the specified organization")
-    @ApiResponses({
-            @ApiResponse(code = 200, message = "Email successfully fetched"),
-            @ApiResponse(code = 500, message = "Internal server error")})
-    public void get(
-            @PathParam("organizationId") String organizationId,
-            @PathParam("environmentId") String environmentId,
-            @PathParam("domain") String domain,
-            @PathParam("client") String client,
-            @NotNull @QueryParam("template") Template emailTemplate,
-            @Suspended final AsyncResponse response) {
-
-        checkAnyPermission(organizationId, environmentId, domain, client, Permission.APPLICATION_EMAIL_TEMPLATE, Acl.READ)
-                .andThen(emailTemplateService.findByDomainAndClientAndTemplate(domain, client, emailTemplate.template())
-                        .map(email -> Response.ok(email).build())
-                        .defaultIfEmpty(Response.status(HttpStatusCode.NOT_FOUND_404).build()))
-                .subscribe(response::resume, response::resume);
-    }
-
-    @POST
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "Create a email for a client",
-            notes = "User must have APPLICATION_EMAIL_TEMPLATE[CREATE] permission on the specified client " +
-                    "or APPLICATION_EMAIL_TEMPLATE[CREATE] permission on the specified domain " +
-                    "or APPLICATION_EMAIL_TEMPLATE[CREATE] permission on the specified environment " +
-                    "or APPLICATION_EMAIL_TEMPLATE[CREATE] permission on the specified organization")
-    @ApiResponses({
-            @ApiResponse(code = 201, message = "Email successfully created"),
-            @ApiResponse(code = 500, message = "Internal server error")})
-    public void create(
-            @PathParam("organizationId") String organizationId,
-            @PathParam("environmentId") String environmentId,
-            @PathParam("domain") String domain,
-            @PathParam("client") String client,
-            @ApiParam(name = "email", required = true)
-            @Valid @NotNull final NewEmail newEmail,
-            @Suspended final AsyncResponse response) {
-        final User authenticatedUser = getAuthenticatedUser();
-
-        checkAnyPermission(organizationId, environmentId, domain, client, Permission.APPLICATION_EMAIL_TEMPLATE, Acl.CREATE)
-                .andThen(domainService.findById(domain)
-                        .switchIfEmpty(Maybe.error(new DomainNotFoundException(domain)))
-                        .flatMap(irrelevant -> clientService.findById(client))
-                        .switchIfEmpty(Maybe.error(new ClientNotFoundException(client)))
-                        .flatMapSingle(__ -> emailTemplateService.create(domain, client, newEmail, authenticatedUser))
-                        .map(email -> Response
-                                .created(URI.create("/organizations/" + organizationId + "/environments/" + environmentId + "/domains/" + domain + "/clients/" + client + "/emails/" + email.getId()))
-                                .entity(email)
-                                .build()))
-                .subscribe(response::resume, response::resume);
-    }
-
-    @Path("{email}")
-    public ClientEmailResource getEmailResource() {
-        return resourceContext.getResource(ClientEmailResource.class);
-    }
-}
