+++ gravitee-am-management-api/gravitee-am-management-api-service/src/test/java/io/gravitee/am/management/service/UserServiceTest.java
@@ -100,9 +100,6 @@ public class UserServiceTest {
     @Mock
     private MembershipService membershipService;
 
-    @Mock
-    private ClientService clientService;
-
     @Spy
     private UserValidator userValidator = new UserValidator();
