--- gravitee-am-repository/gravitee-am-repository-jdbc/src/main/java/io/gravitee/am/repository/jdbc/management/api/JdbcClientRepository.java
@@ -1,49 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.am.repository.jdbc.management.api;
-
-import io.gravitee.am.model.oidc.Client;
-import io.gravitee.am.repository.jdbc.management.AbstractJdbcRepository;
-import io.gravitee.am.repository.management.api.ClientRepository;
-import io.reactivex.Completable;
-import io.reactivex.Single;
-import org.springframework.stereotype.Repository;
-
-import java.util.Collections;
-import java.util.Set;
-
-/**
- * @author Eric LELEU (eric.leleu at graviteesource.com)
- * @author GraviteeSource Team
- */
-@Repository
-public class JdbcClientRepository extends AbstractJdbcRepository implements ClientRepository {
-    // Only used in ClientsToApplicationsUpgrader, useless in this case because the JdbcRepositories will use Applications
-    @Override
-    public Single<Set<Client>> findAll() {
-        return Single.just(Collections.emptySet());
-    }
-
-    @Override
-    public Single<Boolean> collectionExists() {
-        return Single.just(false);
-    }
-
-    @Override
-    public Completable deleteCollection() {
-        return Completable.complete();
-    }
-}
