--- gravitee-am-service/src/main/java/io/gravitee/am/service/model/TotalClient.java
@@ -1,40 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.am.service.model;
-
-/**
- * @author Titouan COMPIEGNE (titouan.compiegne at graviteesource.com)
- * @author GraviteeSource Team
- */
-@Deprecated
-public class TotalClient {
-
-    private long totalClients;
-
-    public TotalClient() { }
-
-    public TotalClient(long totalClients) {
-        this.totalClients = totalClients;
-    }
-
-    public long getTotalClients() {
-        return totalClients;
-    }
-
-    public void setTotalClients(long totalClients) {
-        this.totalClients = totalClients;
-    }
-}
