--- gravitee-am-service/src/main/java/io/gravitee/am/service/model/UpdateClient.java
@@ -1,168 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.am.service.model;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-/**
- * @author David BRASSELY (david.brassely at graviteesource.com)
- * @author GraviteeSource Team
- * @deprecated Please use PatchClient instead.
- */
-public class UpdateClient {
-
-    private List<String> redirectUris;
-
-    private List<String> authorizedGrantTypes;
-
-    private String clientName;
-
-    private List<String> scopes;
-
-    private List<String> autoApproveScopes;
-
-    private int accessTokenValiditySeconds;
-
-    private int refreshTokenValiditySeconds;
-
-    private int idTokenValiditySeconds;
-
-    private Map<String, Object> idTokenCustomClaims;
-
-    private boolean enabled;
-
-    private Set<String> identities;
-
-    private String certificate;
-
-    private boolean enhanceScopesWithUserPermissions;
-
-    private Map<String, Integer> scopeApprovals;
-
-    public List<String> getRedirectUris() {
-        return redirectUris;
-    }
-
-    public void setRedirectUris(List<String> redirectUris) {
-        this.redirectUris = redirectUris;
-    }
-
-    public List<String> getAuthorizedGrantTypes() {
-        return authorizedGrantTypes;
-    }
-
-    public void setAuthorizedGrantTypes(List<String> authorizedGrantTypes) {
-        this.authorizedGrantTypes = authorizedGrantTypes;
-    }
-
-    public String getClientName() {
-        return clientName;
-    }
-
-    public void setClientName(String clientName) {
-        this.clientName = clientName;
-    }
-
-    public List<String> getScopes() {
-        return scopes;
-    }
-
-    public void setScopes(List<String> scopes) {
-        this.scopes = scopes;
-    }
-
-    public List<String> getAutoApproveScopes() {
-        return autoApproveScopes;
-    }
-
-    public void setAutoApproveScopes(List<String> autoApproveScopes) {
-        this.autoApproveScopes = autoApproveScopes;
-    }
-
-    public int getAccessTokenValiditySeconds() {
-        return accessTokenValiditySeconds;
-    }
-
-    public void setAccessTokenValiditySeconds(int accessTokenValiditySeconds) {
-        this.accessTokenValiditySeconds = accessTokenValiditySeconds;
-    }
-
-    public int getRefreshTokenValiditySeconds() {
-        return refreshTokenValiditySeconds;
-    }
-
-    public void setRefreshTokenValiditySeconds(int refreshTokenValiditySeconds) {
-        this.refreshTokenValiditySeconds = refreshTokenValiditySeconds;
-    }
-
-    public boolean isEnabled() {
-        return enabled;
-    }
-
-    public void setEnabled(boolean enabled) {
-        this.enabled = enabled;
-    }
-
-    public Set<String> getIdentities() {
-        return identities;
-    }
-
-    public void setIdentities(Set<String> identities) {
-        this.identities = identities;
-    }
-
-    public int getIdTokenValiditySeconds() {
-        return idTokenValiditySeconds;
-    }
-
-    public void setIdTokenValiditySeconds(int idTokenValiditySeconds) {
-        this.idTokenValiditySeconds = idTokenValiditySeconds;
-    }
-
-    public Map<String, Object> getIdTokenCustomClaims() {
-        return idTokenCustomClaims;
-    }
-
-    public void setIdTokenCustomClaims(Map<String, Object> idTokenCustomClaims) {
-        this.idTokenCustomClaims = idTokenCustomClaims;
-    }
-
-    public String getCertificate() {
-        return certificate;
-    }
-
-    public void setCertificate(String certificate) {
-        this.certificate = certificate;
-    }
-
-    public boolean isEnhanceScopesWithUserPermissions() {
-        return enhanceScopesWithUserPermissions;
-    }
-
-    public void setEnhanceScopesWithUserPermissions(boolean enhanceScopesWithUserPermissions) {
-        this.enhanceScopesWithUserPermissions = enhanceScopesWithUserPermissions;
-    }
-
-    public Map<String, Integer> getScopeApprovals() {
-        return scopeApprovals;
-    }
-
-    public void setScopeApprovals(Map<String, Integer> scopeApprovals) {
-        this.scopeApprovals = scopeApprovals;
-    }
-}
